wit_bindgen_rt

Function maybe_link_cabi_realloc

pub fn maybe_link_cabi_realloc()

This function is called from generated bindings and will be deleted by the linker. The purpose of this function is to force a reference to the symbol cabi_realloc to make its way through to the final linker command line. That way wasm-ld will pick it up, see it needs to be exported, and then export it.

For more information about this see ./ci/rebuild-libcabi-realloc.sh.